+++ portal/helpers/ratelimit.py
@@ -36,22 +36,33 @@ from ratelimit.core import (
     _make_cache_key,
 )
 
+from portal.helpers.regexes import EMAIL_REGEX_PATTERN
+from portal.helpers.request_handlers import get_access_code_from_request
 
 RATELIMIT_LOGIN_GROUP = "login"
 RATELIMIT_LOGIN_RATE = "5/d"
+RATELIMIT_LOGIN_RATE_SCHOOL_STUDENT = "10/d"
 RATELIMIT_METHOD = "POST"
 
 RATELIMIT_USER_ALREADY_REGISTERED_EMAIL_GROUP = "user_already_registered_email"
 RATELIMIT_USER_ALREADY_REGISTERED_EMAIL_RATE = "1/d"
 
 
+def school_student_key(group, request):
+    access_code = get_access_code_from_request(request)
+    return f'{request.POST.get("username", "")},{access_code}'
+
+
 def get_ratelimit_cache_key_for_user(user: str):
-    _, period = _split_rate(rate=RATELIMIT_LOGIN_RATE)
+    # check for email quickly
+
+    user_rate = RATELIMIT_LOGIN_RATE if EMAIL_REGEX_PATTERN.match(user) else RATELIMIT_LOGIN_RATE_SCHOOL_STUDENT
+    _, period = _split_rate(rate=user_rate)
     window = _get_window(value=user, period=period)
     cache_key = _make_cache_key(
         group=RATELIMIT_LOGIN_GROUP,
         window=window,
-        rate=RATELIMIT_LOGIN_RATE,
+        rate=user_rate,
         value=user,
         methods=RATELIMIT_METHOD,
     )
@@ -68,9 +79,7 @@ def clear_ratelimit_cache_for_user(user: str):
     cache.delete(cache_key)
 
 
-def is_ratelimited(
-    request, group=None, fn=None, key=None, rate=None, method=ALL, increment=False
-):
+def is_ratelimited(request, group=None, fn=None, key=None, rate=None, method=ALL, increment=False):
     """
     As in django-ratelimit. Calls "get_usage" defined below to enable the usage of
     the custom cache_key functionality.
@@ -82,9 +91,7 @@ def is_ratelimited(
     return usage["should_limit"]
 
 
-def get_usage(
-    request, group=None, fn=None, key=None, rate=None, method=ALL, increment=False
-):
+def get_usage(request, group=None, fn=None, key=None, rate=None, method=ALL, increment=False):
     """
     As in django-ratelimit. Makes cache_key global so it can be called outside the scope
     and the cache can be accessed at later times.
@@ -92,9 +99,7 @@ def get_usage(
     global cache_key
 
     if group is None and fn is None:
-        raise ImproperlyConfigured(
-            "get_usage must be called with either " "`group` or `fn` arguments"
-        )
+        raise ImproperlyConfigured("get_usage must be called with either " "`group` or `fn` arguments")
 
     if not getattr(settings, "RATELIMIT_ENABLE", True):
         return None