+++ portal/tests/test_daily_activities.py
@@ -1,11 +1,10 @@
+from datetime import timedelta, datetime
+
+from common.models import DailyActivity
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.support.ui import WebDriverWait
 
-
-from common.models import DailyActivity
-
-from datetime import timedelta, datetime
 from portal.tests.base_test import BaseTest